UPDATED: credits were moved to a different file
