+++ cv_next/lib/credits.ts
@@ -1,0 +1,126 @@
+export const Team_Credits = [
+  {
+    name: "Ron Peer",
+    title: "Full Stack Developer",
+    contribution: "Working on the frontend of the project.",
+    linkedIn: "https://www.linkedin.com/in/ron-peer-53b220270/",
+    gitHub: "https://github.com/PeroG28",
+  },
+  {
+    name: "Arad Donenfeld",
+    title: "Security Researcher",
+    contribution: "Leading the development of the project.",
+    linkedIn: "https://www.linkedin.com/in/aradon267/",
+    gitHub: "https://github.com/Aradon267",
+  },
+  {
+    name: "Denis Granovsky",
+    title: "Software Developer",
+    contribution: "Managing the development of the project.",
+    linkedIn: "https://www.linkedin.com/in/denisgran/",
+    gitHub: "https://github.com/DenisGran",
+  },
+  {
+    name: "Yarden Perets",
+    title: "Software Engineer",
+    contribution: "Working on development and DB stability for the poject.",
+    linkedIn: "https://www.linkedin.com/in/yarden-perets/",
+    gitHub: "https://github.com/lordYorden",
+  },
+  {
+    name: "Assaf Kabesa",
+    title: "Software Developer",
+    contribution: "Working on the deployment process and cloud services.",
+    linkedIn: "https://www.linkedin.com/in/assaf-kabesa/",
+    gitHub: "https://github.com/AssafLolGG",
+  },
+  {
+    name: "Omri Anidgar",
+    title: "Full Stack Developer",
+    contribution:
+      "Working on the deployment process and frontend of the project.",
+    linkedIn: "https://www.linkedin.com/in/omri-anidgar/",
+    gitHub: "https://github.com/Omria09",
+  },
+  {
+    name: "Shoham Yosef Bitton",
+    title: "Developer",
+    contribution:
+      "Working on the deployment process and automations for the project.",
+    linkedIn: "https://www.linkedin.com/in/shohamyosefbitton/",
+    gitHub: "https://github.com/ShohamBit",
+  },
+  {
+    name: "Alon Green",
+    title: "Cloud Security Engineer",
+    contribution: "Working on the frontend of the project.",
+    linkedIn: "https://www.linkedin.com/in/assaf-kabesa/",
+    gitHub: "https://github.com/TheGreenOak",
+  },
+  {
+    name: "Nevo Sznajder",
+    title: "Full Stack Developer",
+    contribution: "Working on backend development of the project.",
+    linkedIn: "https://www.linkedin.com/in/nevo-sznajder-916622279/",
+    gitHub: "https://github.com/nevogithub",
+  },
+  {
+    name: "Yechiam Weiss",
+    title: "Full Stack Developer",
+    contribution: "Advising us through the full product development process.",
+    linkedIn: "https://www.linkedin.com/in/yechiamweiss/",
+    gitHub: "https://github.com/YechiamW3",
+  },
+];
+
+export const External_Credits = [
+  {
+    name: "Adam Abramov",
+    title: "Tech Lead Reverse Engineer",
+    avatar_url: "",
+    contribution:
+      "Insufficient Verification of Data Authenticity & CSRF Account Takeover.",
+    linkedIn: "https://www.linkedin.com/in/theresearcher/",
+  },
+  {
+    name: "Youval Daizi",
+    title: "Cloud Security",
+    contribution: "Home button caused users to log out.",
+    linkedIn: "https://www.linkedin.com/in/youval-daizi/",
+  },
+  {
+    name: "Harel Kristal",
+    title: "Penetration Tester",
+    contribution:
+      "Login and Signout were always visible & Developer's data leak.",
+    linkedIn: "https://www.linkedin.com/in/harel-kristal/",
+  },
+  {
+    name: "Avner Mindelis",
+    title: "Security Researcher",
+    contribution: "Lead initial development of the project.",
+    linkedIn: "https://www.linkedin.com/in/avner-mindelis/",
+    github: "https://github.com/avnogy",
+  },
+  {
+    name: "Adam Liberov",
+    title: "Full Stack Developer",
+    contribution: "Worked on initial development of the project.",
+    linkedIn: "https://www.linkedin.com/in/adam-liberov-9918061b3/",
+    github: "https://github.com/AdamLiberov",
+  },
+  {
+    name: "Ethan Krimer",
+    title: "Full Stack Developer",
+    contribution: "Managed initial development of the project.",
+    linkedIn: "https://www.linkedin.com/in/ethankrimer/",
+  },
+  {
+    name: "Ilan Yashuk",
+    title: "",
+    contribution: "Worked on initial development of the project.",
+    linkedIn: "https://www.linkedin.com/in/ilan-yashuk/",
+    gitHub: "https://github.com/ya5huk",
+  },
+  // Add more contributors as needed
+];
